Adds Null Check to Remove Configuration.
Adds a null check to the stats manager android framework code for remove
configuration to prevent a potential null pointer.

- this crash happens 5k-11k times per day
- root cause unknown

Test: presubmit
Bug: 431099361
Flag: EXEMPT bugfix

+++ framework/java/android/app/StatsManager.java
@@ -217,6 +217,10 @@ public final class StatsManager {
         synchronized (sLock) {
             try {
                 IStatsManagerService service = getIStatsManagerServiceLocked();
+                if (service == null) {
+                    Log.e(TAG, "StatsManagerService is null when removing configuration");
+                    throw new StatsUnavailableException("could not connect");
+                }
                 service.removeConfiguration(configKey, mContext.getOpPackageName());
             } catch (RemoteException e) {
                 Log.e(TAG, "Failed to connect to statsmanager when removing configuration");
